Added an extra check to avoid exception
@chriswmackey I had to add an extra check to make sure this won't crash
Honeybee_Honeybee

+++ src/Honeybee_Honeybee.py
@@ -485,7 +485,7 @@ class PrepareTemplateEPLibFiles(object):
             libFilePaths.append(customEPLib)
         
         #download THERM template file.
-        if sc.sticky["isNewerTHERMAvailable"] or not os.path.isfile(thermTemplateFile):
+        if sc.sticky.has_key("isNewerTHERMAvailable") and sc.sticky["isNewerTHERMAvailable"] or not os.path.isfile(thermTemplateFile):
             # create a backup from users library
             try: shutil.copyfile(thermTemplateFile, thermBckupfile)
             except: pass
